Add pre_processor functionality to feature_extraction API
Currently if feature extractor x depends on feature extractor y, feature y is processed in the process method of x, which is a bit messy. Also, the unit of y has to be set in an overidden set_unit function in x, which I might forget to implement (that would lead to wrong feature results). With the new API both issues are addressed.